+++ spec/unit/optimizer/test_allbut.rb
@@ -9,6 +9,12 @@ module Bmg
       ])
     }
 
+    before do
+      class Operator::Page
+        public :ordering, :page_index, :options
+      end
+    end
+
     context 'allbut on empty butlist' do
       subject{
         relation.allbut([])
@@ -16,6 +22,53 @@ module Bmg
 
       it 'returns the operand itself' do
         expect(subject).to be(relation)
+      end
+    end
+
+    context 'allbut.page' do
+      subject{
+        relation.allbut(butlist).page(ordering, 1, :page_size => 10)
+      }
+
+      context 'when keys are not known' do
+        let(:butlist) {
+          [:c]
+        }
+        let(:ordering) {
+          [[:a, :asc]]
+        }
+
+        it 'is not optimized since duplicate removal changes the pages' do
+          expect(subject).to be_a(Operator::Page)
+          expect(operand).to be_a(Operator::Allbut)
+          expect(operand(operand)).to be(relation)
+        end
+      end
+
+
+      context 'when a key is preserved' do
+        let(:relation) {
+          Relation.new([
+            { a: 1,  b: 2, c: 3 },
+            { a: 11, b: 2, c: 33 }
+          ]).with_type(Type::ANY.with_attrlist([:a, :b, :b]).with_keys([[:a]]))
+        }
+        let(:butlist) {
+          [:c]
+        }
+        let(:ordering) {
+          [[:a, :asc]]
+        }
+
+        it 'the page is pushed down the tree' do
+          expect(subject).to be_a(Operator::Allbut)
+          expect(subject.butlist).to eql(butlist)
+          expect(operand).to be_a(Operator::Page)
+          expect(operand.ordering).to eql(ordering)
+          expect(operand.page_index).to eql(1)
+          expect(operand.options).to eql(:page_size => 10)
+          expect(operand(operand)).to be(relation)
+        end
       end
     end