+++ CHANGELOG.md
@@ -27,6 +27,9 @@
 * Optimize `transform.restrict`. Push whatever can be pushed down the
   tree.
 
+* Optimize `allbut.page`. Push the page down the tree if allbut is
+  key preserving.
+
 # 0.18.0 - 2021/03/12
 
 * Default Relation#type is provided, that returns Bmg::Type::ANY